+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `Modal` component.
 - Feature: Add `withoutTopBottomBorder`, `borderColor` and other props
   to `InfoLines` component.
 - Feature: `LinkList` responsive.